refactor(tools): split formatter tool from lsp
Move the formatter tool into its own module so LSP remains focused on language-server lookups while preserving the existing format_file contract.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
